third_party/WebKit/LayoutTests/http/tests/svg/svg-in-object-in-sandboxed-iframe.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test("SVG in <object> in sandboxed <iframe> is not blocked.");
onmessage = t.step_func_done(function(message) {
  assert_equals(message.data, 'load');
  var iframe = document.querySelector('iframe');
  var containedObject = iframe.contentDocument.querySelector('object');
  var svgRoot = containedObject.contentDocument.documentElement;
  assert_equals(svgRoot.localName, "svg");
  assert_equals(svgRoot.namespaceURI, "http://www.w3.org/2000/svg");
});
</script>
<iframe sandbox="allow-same-origin allow-scripts" src="resources/object-svg.html"
        style="border: none; margin: 0"
        scrolling="no"></iframe>
